Change to using rust modules instead of sub-crates
Decided it would be better to just use the more common usage of rust
modules instead of having a bunch of local crates. This will also allow
for a test directory in the future to start some tests.
