+++ Cargo.toml
@@ -7,28 +7,6 @@ authors = ["Henry Estela <henryestela@gmail.com>"]
 [lib]
 name = "rust_psm"
 
-# PSM Endpoint
-[dependencies.ep]
-path = "./ep"
-
-[dependencies.error]
-path = "./error"
-
-# Packet Transport Layer - Active Messages
-[dependencies.ptl_am]
-path = "./ptl_am"
-
-# Packet Transport Layer - Interconnect Protocol Stack
-[dependencies.ptl_ips]
-path = "./ptl_ips"
-
-# Packet Transport Layer - Self (Loopback)
-[dependencies.ptl_self]
-path = "./ptl_self"
-
-# Infinipath
-[dependencies.ipath]
-path = "./ipath"
-
 [dependencies]
-uuid = "0.1"
+libc = "0.2"
+uuid = "0.2"
