--- ep/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-
-name = "ep"
-version = "0.0.1"
-authors = ["Henry Estela <henryestela@gmail.com>"]
-
-[dependencies.error]
-path = "./../error"
-
-[dependencies]
-uuid = "0.1"
