--- ptl_am/Cargo.toml
@@ -1,5 +1,0 @@
-[package]
-
-name = "ptl_am"
-version = "0.0.1"
-authors = ["Henry Estela <henryestela@gmail.com>"]
